rename screens for better clarification on usage, add event list to show automatically on user dashboard
